% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_ETG4000.R
\name{epoch_ETG4000}
\alias{epoch_ETG4000}
\title{Epoch the ETG-4000 data}
\usage{
epoch_ETG4000(x, trigger, start_epoch, end_epoch)
}
\arguments{
\item{x}{A data frame with an ETG-4000 data.}

\item{trigger}{A code used to represent a trigger, e.g. 1.}

\item{start_epoch}{A numeric. Number of samples before the trigger. Start of the epoch.}

\item{end_epoch}{A numeric. Number of samples after the trigger. End of the epoch.}
}
\value{
A data frame.
}
\description{
Epoch the ETG-4000 data
}
\examples{
file_path <- system.file("extdata", "Hitachi_ETG4000_24Ch_Total.csv", package = "fnirsr")
ETG_header <- load_ETG4000_header(file_path)
rawData <- load_ETG4000_data(file_path, ETG_header)

df_epochs <- epoch_ETG4000(rawData, 1, -100, 100)

}
