% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_ETG4000_average.R
\name{epoch_ETG4000_average}
\alias{epoch_ETG4000_average}
\title{Average the epoched ETG-4000 data}
\usage{
epoch_ETG4000_average(x, start_epoch, end_epoch)
}
\arguments{
\item{x}{A data frame with an ETG-4000 data.}

\item{start_epoch}{A numeric. Number of samples before the trigger. Start of the epoch. The same as in epoch_ETG4000().}

\item{end_epoch}{A numeric. Number of samples after the trigger. End of the epoch. The same as in epoch_ETG4000().}
}
\value{
A data frame.
}
\description{
Average the epoched ETG-4000 data
}
\examples{
file_path <- system.file("extdata", "Hitachi_ETG4000_24Ch_Total.csv", package = "fnirsr")
ETG_header <- load_ETG4000_header(file_path)
rawData <- load_ETG4000_data(file_path, ETG_header)
df_epochs <- epoch_ETG4000(rawData, 1, -100, 100) # use to explore the epoch quality
df_epochs_average <- epoch_ETG4000_average(df_epochs, -100, 100)

}
