% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grand_average_ETG4000.R
\name{grand_average_ETG4000}
\alias{grand_average_ETG4000}
\title{Create a grand average of the ETG-4000 data}
\usage{
grand_average_ETG4000(x)
}
\arguments{
\item{x}{A data frame with ETG-4000 data.}
}
\value{
A data frame with ETG-4000 data plus an extra column including Grand Average
}
\description{
Create a grand average of the ETG-4000 data
}
\examples{
file_path <- system.file("extdata", "Hitachi_ETG4000_24Ch_Total.csv", package = "fnirsr")
ETG_header <- load_ETG4000_header(file_path)

fnirs_signal <- load_ETG4000_data(file_path, ETG_header)
grand_average_ETG4000(fnirs_signal)

}
