% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitachi_to_nirs.R
\name{hitachi_to_nirs}
\alias{hitachi_to_nirs}
\title{Load the data from a nirs file}
\usage{
hitachi_to_nirs(x)
}
\arguments{
\item{x}{A Hitachi fNIRS file.}
}
\value{
A .nirs file.
}
\description{
Load the data from a nirs file
}
\examples{
hitachi_to_nirs("....csv")

}
\references{
\url{https://www.nitrc.org/projects/hitachi2nirs/}
}

