% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ETG4000_data.R
\name{load_ETG4000_data}
\alias{load_ETG4000_data}
\title{Load the ETG-4000 data}
\usage{
load_ETG4000_data(x, header_ETG4000)
}
\arguments{
\item{x}{Raw csv file.}

\item{header_ETG4000}{A list representing an ETG-4000 header.}
}
\value{
A data frame.
}
\description{
Load the ETG-4000 data
}
\examples{
file_path <- system.file("extdata", "Hitachi_ETG4000_24Ch_Total.csv", package = "fnirsr")
ETG_header <- load_ETG4000_header(file_path)
load_ETG4000_data(file_path, ETG_header)

}
