% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ETG4000_header.R
\name{load_ETG4000_header}
\alias{load_ETG4000_header}
\title{Load the ETG-4000 data}
\usage{
load_ETG4000_header(x)
}
\arguments{
\item{x}{Raw csv file.}
}
\value{
A vector.
}
\description{
Load the ETG-4000 data
}
\examples{
file_path <- system.file("extdata", "Hitachi_ETG4000_24Ch_Total.csv", package = "fnirsr")
load_ETG4000_header(file_path)

}
