% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nirs_data.R
\name{load_nirs_data}
\alias{load_nirs_data}
\title{Load the data from a nirs file}
\usage{
load_nirs_data(x)
}
\arguments{
\item{x}{A nirs file.}
}
\value{
A list.
}
\description{
Load the data from a nirs file
}
\examples{
file_path <- system.file("extdata", "Simple_Probe.nirs", package = "fnirsr")
load_nirs_data(file_path)

}
