% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ETG4000.R
\name{plot_ETG4000}
\alias{plot_ETG4000}
\title{Plot the raw signal}
\usage{
plot_ETG4000(x, type = "facets", channel = NULL)
}
\arguments{
\item{x}{A data frame with an ETG-4000 data.}

\item{type}{A string. Specifies the type of plot.
Options: "facets", "overlap", "separate", "average". Defaults to "facets".}

\item{channel}{An integer. Represents the channel number.}
}
\value{
A plot.
}
\description{
Plot the raw signal
}
\examples{
file_path <- system.file("extdata", "Hitachi_ETG4000_24Ch_Total.csv", package = "fnirsr")
ETG_header <- load_ETG4000_header(file_path)
rawData <- load_ETG4000_data(file_path, ETG_header)

plot_ETG4000(rawData)
plot_ETG4000(rawData, type = "overlap")
plot_ETG4000(rawData, type = "separate", channel = 1)

averagedData <- grand_average_ETG4000(rawData)
plot_ETG4000(averagedData, type = "average")

}
