% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nirs.R
\name{plot_nirs}
\alias{plot_nirs}
\title{Plot the data extracted from a nirs file}
\usage{
plot_nirs(x, type = "facets")
}
\arguments{
\item{x}{A list.}
}
\value{
A plot.
}
\description{
Plot the data extracted from a nirs file
}
\examples{
data <- load_nirs_data("Simple_Probe.nirs")
plot_nirs(data)
plot_nirs(data, type = "facets")

}

