% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_channels_ETG4000.R
\name{remove_channels_ETG4000}
\alias{remove_channels_ETG4000}
\title{Remove channels from the ETG-4000 data}
\usage{
remove_channels_ETG4000(x, channel = NULL)
}
\arguments{
\item{x}{A data frame with ETG-4000 data.}

\item{channel}{A vector with integers.}
}
\value{
A data frame with ETG-4000 data.
}
\description{
Remove channels from the ETG-4000 data
}
\examples{
grand_average_ETG4000(df, channel = 15)
grand_average_ETG4000(df, channel = c(15, 20))

}

