% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_triggers.R
\name{remove_triggers}
\alias{remove_triggers}
\title{Remove triggers from the ETG-4000 data}
\usage{
remove_triggers(x, trigger = NULL)
}
\arguments{
\item{x}{A data frame with an ETG-4000 data.}

\item{trigger}{An integer representign a trigger/marking to remove.}
}
\value{
A data frame.
}
\description{
Remove triggers from the ETG-4000 data
}
\examples{
file_path <- system.file("extdata", "Hitachi_ETG4000_24Ch_Total.csv", package = "fnirsr")
ETG_header <- load_ETG4000_header(file_path)
ETG_data <-load_ETG4000_data(file_path, ETG_header)
ETG_data_cleaned <- remove_triggers(ETG_data, 1)

}
