% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_ETG4000.R
\name{detrend_ETG4000_data}
\alias{detrend_ETG4000_data}
\title{Detrend the ETG-4000 data.}
\usage{
detrend_ETG4000_data(x)
}
\arguments{
\item{x}{Raw csv file.}
}
\value{
A data frame.
}
\description{
Removes a linear trend.
}
\examples{
file_path <- system.file("extdata", "Hitachi_ETG4000_24Ch_Total.csv", package = "fnirsr")
fnirs_signal <- load_ETG4000_data(file_path)
plot_ETG4000(fnirs_signal)
fnirs_det <- detrend_ETG4000_data(fnirs_signal)
plot_ETG4000(fnirs_det)

plot_ETG4000(fnirs_signal, "separate", 18)
plot_ETG4000(fnirs_det, "separate", 18)

}

