% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ETG4000_data.R
\name{load_ETG4000_data}
\alias{load_ETG4000_data}
\title{Load the ETG-4000 data}
\usage{
load_ETG4000_data(x)
}
\arguments{
\item{x}{Raw csv file.}
}
\value{
A data frame.
}
\description{
Load the ETG-4000 data
}
\examples{
file_path <- system.file("extdata", "Hitachi_ETG4000_24Ch_Total.csv", package = "fnirsr")
load_ETG4000_data(file_path)

}

