% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_channels_ETG4000.R
\name{remove_channels_ETG4000}
\alias{remove_channels_ETG4000}
\title{Remove channels from the ETG-4000 data}
\usage{
remove_channels_ETG4000(x, channel = NULL)
}
\arguments{
\item{x}{A data frame with ETG-4000 data.}

\item{channel}{A vector with integers.}
}
\value{
A data frame with ETG-4000 data.
}
\description{
Remove channels from the ETG-4000 data
}
\examples{
file_path <- system.file("extdata", "Hitachi_ETG4000_24Ch_Total.csv", package = "fnirsr")
ETG_header <- load_ETG4000_header(file_path)
df <- load_ETG4000_data(file_path, ETG_header)

remove_channels_ETG4000(df, channel = 15)
remove_channels_ETG4000(df, channel = c(16, 20))

}

